requirements.txt
opencv-python>=4.8.0
python-chess>=1.9.0
pygame>=2.5.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.0.0
